fix: extract you config init wiring to root_config.go
Move system config command registration out of root.go so the file stays under the backend-size 1000-line limit flagged in PR review.
